requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
numpy==1.26.0
openpyxl==3.1.2
outcome==1.2.0
pandas==2.1.1
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
wsproto==1.2.0
